@@ -3,18 +3,21 @@
   "version": "4.0.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "latest",
-    "@material-ui/icons": "^4.11.2",
-    "@primer/octicons-react": "^12.0.0",
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.10.6",
+    "@mui/material": "^5.10.8",
+    "@mui/styles": "^5.10.8",
+    "@primer/octicons-react": "^17.6.0",
     "clsx": "latest",
     "next": "latest",
-    "next-pwa": "^5.0.5",
+    "next-pwa": "^5.6.0",
     "prop-types": "latest",
     "react": "latest",
     "react-dom": "latest",
     "react-typed": "^1.2.0",
-    "simple-icons": "^4.11.0",
-    "webpack": "^4.46.0"
+    "simple-icons": "^7.14.0",
+    "webpack": "^5.74.0"
   },
   "scripts": {
     "dev": "next",
@@ -22,8 +25,5 @@
     "start": "next start",
     "post-update": "echo \"codesandbox preview only, need an update\" && yarn upgrade --latest",
     "deploy": "npx vercel --prod"
-  },
-  "devDependencies": {
-    "@types/simple-icons": "^4.10.0"
   }
 }
