@@ -1,16 +1,18 @@
-import { Grid, makeStyles, Typography, Tooltip, Avatar } from "@material-ui/core";
-import data from '../data.json'
-import simpleIcons from 'simple-icons'
+import { Grid, Typography, Tooltip, Avatar } from "@mui/material";
+import { makeStyles } from "@mui/styles";
+import dynamic from 'next/dynamic'
+import { Suspense } from 'react'
+import data from '../data.json';
 import clsx from "clsx";
 import Image from 'next/image'
 import { iconify } from "./util";
-import Cancel from "@material-ui/icons/Cancel";
+import Cancel from "@mui/icons-material/Cancel";
 const { about } = data
 
 const dpx = about.social.length*10 - 2
 
 const socialDetails = about.social.map(({ alt, icon, link }) => {
-    const ic = simpleIcons.get(iconify(icon)) || {
+    const ic = dynamic(() => import('simple-icons/icons').then((mod) => mod[icon])) || {
         hex: '424242',
         component: <Cancel color="white" fontSize={36}/>
     }
@@ -32,7 +34,7 @@ socialDetails.forEach(({ alt, backgroundColor }) => {
 
 const useStyles = makeStyles(theme => ({
     cont: {
-        minHeight: `calc(100vh - ${theme.spacing(4)}px)`,
+        minHeight: `calc(100vh - ${theme.spacing(4)})`,
         alignSelf: 'center',
         justifySelf: 'center'
     },
