@@ -1,21 +1,21 @@
-import { createMuiTheme } from "@material-ui/core";
+import { createTheme } from "@mui/material";
 import data from "../data.json";
 const { theme } = data
 
-export const lightTheme = createMuiTheme(
+export const lightTheme = createTheme(
   (theme && theme.light && Object.keys(theme.light).length > 0) ? theme.light :
   {
     palette: {
-      type: "light"
+      mode: "light"
     }
   }
 )
 
-export const darkTheme = createMuiTheme(
+export const darkTheme = createTheme(
   (theme && theme.dark && Object.keys(theme.dark).length > 0) ? theme.dark :
   {
     palette: {
-      type: "dark"
+      mode: "dark"
     }
   }
 )
