@@ -1,11 +1,16 @@
-const withPWA = require('next-pwa')
-
-module.exports = withPWA({
-    pwa: {
-        disable: process.env.NODE_ENV === 'development',
-        dest: 'public'
-    },
-    images: {
-        domains: ['media-exp1.licdn.com']
-    }
+const withPWA = require('next-pwa')({
+    dest: 'public',
+    disable: process.env.NODE_ENV === 'development',
+    register: true,
+    //scope: '/app',
+    //sw: 'service-worker.js',
+    //...
 })
+  
+  module.exports = withPWA({
+    reactStrictMode: false,
+    swcMinify: true,
+    images: {
+      domains: ['media-exp1.licdn.com']
+    }
+  }) 
